pyproject.toml
[project]
name = "multiexfeed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "pandas>=2.3.3",
    "pandas-ta>=0.4.71b0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
